requirements.txt
aioredis==2.0.0
asgiref==3.4.1
async-timeout==3.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
dnspython==1.16.0
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
loguru==0.5.3
motor==2.5.1
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.1
PyYAML==6.0
requests==2.26.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uuid==1.30
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
